+++ Funcky.Test/ResultTest.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using Funcky.Monads;
 using Xunit;
+using static Funcky.Functional;
 
 namespace Funcky.Test
 {
@@ -13,8 +14,8 @@ namespace Funcky.Test
             var value = Result<int>.Ok(1000);
 
             var hasResult = value.Match(
-                ok: x => true,
-                error: y => false);
+                ok: True,
+                error: False);
 
             Assert.True(hasResult);
         }
@@ -25,8 +26,8 @@ namespace Funcky.Test
             var value = Result<int>.Error(new ArgumentException());
 
             var hasResult = value.Match(
-                ok: x => true,
-                error: y => false);
+                ok: True,
+                error: False);
 
             Assert.False(hasResult);
         }
@@ -79,6 +80,24 @@ namespace Funcky.Test
                 { Result<int>.Ok(1337), Result<int>.Ok(42), Result<int>.Ok(99), 1478 },
                 { Result<int>.Ok(45856), Result<int>.Ok(58788), Result<int>.Ok(699554), 804198 },
                 { Result<int>.Error(new InvalidCastException()), Result<int>.Error(new IOException()), Result<int>.Error(new MemberAccessException()), null },
+            };
+
+        [Theory]
+        [MemberData(nameof(GetResults))]
+        public void MatchAcceptsActionsAsFunctions(Result<int> result, bool expected)
+        {
+            result
+              .Match(
+                ok: v => Assert.True(expected),
+                error: e => Assert.False(expected));
+        }
+
+        public static TheoryData<Result<int>, bool> GetResults()
+            => new TheoryData<Result<int>, bool>
+            {
+                { Result<int>.Ok(5), true },
+                { Result<int>.Ok(42), true },
+                { Result<int>.Error(new InvalidCastException()), false },
             };
     }
 }