+++ Funcky/Extensions/DictionaryExtensions.cs
@@ -6,6 +6,8 @@ namespace Funcky.Extensions
     public static class DictionaryExtensions
     {
         public static Option<TValue> TryGetValue<TKey, TValue>(this IDictionary<TKey, TValue> dictionary, TKey key)
+            where TKey : notnull
+            where TValue : notnull
         {
             return dictionary.TryGetValue(key, out var result)
                 ? Option.Some(result)
@@ -13,6 +15,8 @@ namespace Funcky.Extensions
         }
 
         public static Option<TValue> TryGetValue<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> dictionary, TKey readOnlyKey)
+            where TKey : notnull
+            where TValue : notnull
         {
             return dictionary.TryGetValue(readOnlyKey, out var result)
                 ? Option.Some(result)