+++ Funcky/Functional/BoolConstant.cs
@@ -1,0 +1,9 @@
+﻿namespace Funcky
+{
+    public static partial class Functional
+    {
+        public static bool True<T>(T dummy) => true;
+
+        public static bool False<T>(T dummy) => false;
+    }
+}