+++ Funcky/GenericConstraints/RequireClass.cs
@@ -1,0 +1,7 @@
+namespace Funcky.GenericConstraints
+{
+    public sealed class RequireClass<T>
+        where T : class
+    {
+    }
+}