+++ Funcky/Monads/Option.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Funcky.GenericConstraints;
 
 namespace Funcky.Monads
 {
-    public readonly struct Option<TItem> :
-        IToString
+    public readonly struct Option<TItem> : IToString
+        where TItem : notnull
     {
         private readonly bool _hasItem;
         private readonly TItem _item;
@@ -28,29 +29,15 @@ namespace Funcky.Monads
         public static Option<TItem> None() => default;
 
         public Option<TResult> Select<TResult>(Func<TItem, TResult> selector)
-        {
-            if (selector == null)
-            {
-                throw new ArgumentNullException(nameof(selector));
-            }
-
-            return _hasItem
+            where TResult : notnull
+            => _hasItem
                 ? Option.Some(selector(_item))
                 : Option<TResult>.None();
-        }
 
         public Option<TResult> SelectMany<TMaybe, TResult>(Func<TItem, Option<TMaybe>> maybeSelector, Func<TItem, TMaybe, TResult> resultSelector)
+            where TResult : notnull
+            where TMaybe : notnull
         {
-            if (maybeSelector == null)
-            {
-                throw new ArgumentNullException(nameof(maybeSelector));
-            }
-
-            if (resultSelector == null)
-            {
-                throw new ArgumentNullException(nameof(resultSelector));
-            }
-
             if (_hasItem)
             {
                 var selectedMaybe = maybeSelector(_item);
@@ -64,46 +51,17 @@ namespace Funcky.Monads
         }
 
         public TResult Match<TResult>(TResult none, Func<TItem, TResult> some)
-        {
-            if (some == null)
-            {
-                throw new ArgumentNullException(nameof(some));
-            }
-
-            return _hasItem
+            => _hasItem
                 ? some(_item)
                 : none;
-        }
 
         public TResult Match<TResult>(Func<TResult> none, Func<TItem, TResult> some)
-        {
-            if (none == null)
-            {
-                throw new ArgumentNullException(nameof(none));
-            }
-
-            if (some == null)
-            {
-                throw new ArgumentNullException(nameof(some));
-            }
-
-            return _hasItem
-                ? some(_item)
-                : none();
-        }
+            => _hasItem
+                  ? some(_item)
+                  : none();
 
         public void Match(Action none, Action<TItem> some)
         {
-            if (none == null)
-            {
-                throw new ArgumentNullException(nameof(none));
-            }
-
-            if (some == null)
-            {
-                throw new ArgumentNullException(nameof(some));
-            }
-
             if (_hasItem)
             {
                 some(_item);
@@ -115,39 +73,30 @@ namespace Funcky.Monads
         }
 
         public Option<TItem> OrElse(Option<TItem> elseOption)
-        {
-            return _hasItem
+            => _hasItem
                 ? this
                 : elseOption;
-        }
 
         public TItem OrElse(TItem elseOption)
-        {
-            return _hasItem
+            => _hasItem
                 ? _item
                 : elseOption;
-        }
 
         public Option<TItem> OrElse(Func<Option<TItem>> elseOption)
-        {
-            return _hasItem
+            => _hasItem
                 ? this
                 : elseOption.Invoke();
-        }
 
         public TItem OrElse(Func<TItem> elseOption)
-        {
-            return _hasItem
+            => _hasItem
                 ? _item
                 : elseOption.Invoke();
-        }
 
         public Option<TResult> AndThen<TResult>(Func<TItem, TResult> andThenFunction)
-        {
-            return _hasItem
+            where TResult : notnull
+            => _hasItem
                 ? Option.Some(andThenFunction(_item))
                 : Option<TResult>.None();
-        }
 
         public void AndThen(Action<TItem> andThenFunction)
         {
@@ -167,36 +116,43 @@ namespace Funcky.Monads
                    some: value => Enumerable.Repeat(value, 1));
 
         public override bool Equals(object obj)
-        {
-            return obj is Option<TItem> other
-                   && Equals(_item, other._item);
-        }
+            => obj is Option<TItem> other
+            && Equals(_item, other._item);
 
-        public override int GetHashCode()
-        {
-            return _hasItem
-                ? _item.GetHashCode()
-                : 0;
-        }
+        public override int GetHashCode() =>
+            Select(item => item.GetHashCode()).OrElse(0);
 
         public override string ToString()
-        {
-            return Match(
+            => Match(
                 none: "None",
                 some: value => $"Some({value})");
-        }
     }
 
     public static class Option
     {
         public static Option<TItem> Some<TItem>(TItem item)
-        {
-            return new Option<TItem>(item);
-        }
+            where TItem : notnull
+            => new Option<TItem>(item);
 
         public static Option<TItem> Some<TItem>(Option<TItem> item)
-        {
-            return item;
-        }
+            where TItem : notnull
+            => item;
+
+        /// <summary>
+        /// Creates an <see cref="Option{T}"/> from a nullable value.
+        /// </summary>
+        public static Option<T> From<T>(T? item, RequireClass<T>? ω = null)
+            where T : class
+            => item is { } value ? Some(value) : Option<T>.None();
+
+        /// <inheritdoc cref="From{T}(T, RequireClass{T})"/>
+        public static Option<T> From<T>(T item, RequireStruct<T>? ω = null)
+            where T : struct
+            => Some(item);
+
+        /// <inheritdoc cref="From{T}(T, RequireClass{T})"/>
+        public static Option<T> From<T>(T? item)
+            where T : struct
+            => item.HasValue ? Some(item.Value) : Option<T>.None();
     }
 }