+++ changelog.md
@@ -5,7 +5,15 @@
 * Add `WhereSelect` extension function for `IEnumerable<T>`.
 * Add missing overload for nullary actions to `ActionToUnit`.
 
-## 2.0.0
+## 1.7.0
+* Add nullability annotations to everything except for `Monads.Reader`.
+* Add a function for creating an `Option<T>` from a nullable value: `Option.From`.
+* `Either.Match` now throws when called on an `Either` value created using `default(Either<L, R>)`.
+* Add `True` and `False` functions to public API
+* Match of `Result` Monad accepts actions
+* Add `FirstOrNone`, `LastOrNone` and `SingleOrNone` extension functions
+
+## Next major version
 * Use `Func<T, bool>` instead of `Predicate<T>` in predicate composition functions (`Functional.All`, `Functional.Any`, `Functional.Not`),
   because most APIs in `System` use `Func`.
 * `Functional.Any` now returns `false` when the given list of predicates is empty.